requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.2.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.7
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.3.2
chromadb==1.0.12
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
cryptography==44.0.0
curl_cffi==0.12.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
docutils==0.21.2
docx2txt==0.8
duckduckgo_search==8.0.4
durationpy==0.10
executing==2.2.0
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.53.1
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.70.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.9.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
kubernetes==33.1.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
language_data==1.3.0
libclang==18.1.1
lightgbm==4.6.0
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mmh3==5.1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.12
murmurhash==1.0.12
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.32.0
networkx==3.5
nltk==3.9.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.86.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.18.2
phonenumbers==8.13.53
pillow==10.4.0
platformdirs==4.3.8
plotly==6.0.1
posthog==4.10.0
preshed==3.0.9
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.3
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydeck==0.9.1
pydotplus==2.0.2
pyfiglet==1.0.2
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyInputPlus==0.2.12
pylance==0.9.7
PyMsgBox==1.0.9
PyMuPDF==1.26.1
pyparsing==3.1.2
pypdfium2==4.30.1
pyperclip==1.9.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySimpleValidate==0.2.12
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resume-parser==0.8.4
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.34.2
sentence-transformers==4.1.0
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
SKompiler==0.7
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
stdiomask==0.0.6
stemming==1.0.1
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
thinc==8.3.4
threadpoolctl==3.5.0
tika==2.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.0
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.65
zipp==3.23.0
zstandard==0.23.0
